fix CI

+++ src/zmq/zmqpublishnotifier.cpp
@@ -373,7 +373,7 @@ bool CZMQPublishNEVMBlockConnectNotifier::NotifyNEVMBlockConnect(const CNEVMHead
     LogPrint(BCLog::ZMQ, "zmq: Publish nevm block connect %s to %s, subscriber %s\n", hash.GetHex(), this->address, this->addresssub);
     CDataStream ss(SER_NETWORK, PROTOCOL_VERSION);
     ss << evmBlock << block.vchNEVMBlockData << nSYSBlockHash;
-    if(nHeight >= Params().GetConsensus().nPODAStartBlock)
+    if(nHeight >= (uint32_t)Params().GetConsensus().nPODAStartBlock)
         ss << NEVMDataVecOut;
     if(!SendZmqMessageNEVM(MSG_NEVMBLOCKCONNECT, &(*ss.begin()), ss.size()))
         return state.Invalid(BlockValidationResult::BLOCK_INVALID_HEADER, "nevm-connect-not-sent");
